default arch to x86_64
The default arch of Arch Linux is x86_64 so support that out of the box.
To build anything using cargo-arch, an arch is required but it was
implemented as an Option which doesn't make any sense if its required.
Clear this up by always defaulting to x86_64 if the user hasn't
configured this in Cargo.toml but allow them to override it. fixes #6